# Fase-I/guidelines/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sistema de Análisis de Sentimiento de Reviews – Hotel OLH",
    "description": "API REST para el pipeline de NLP y el dashboard ejecutivo de análisis de sentimiento de reseñas hoteleras. Cubre los servicios: Data Verification & Cleaner, Sentiment Prediction, ABSAService (ABSA sobre tópicos hoteleros → `review_topicos`), Topic Identification (alternativa al ABSA para la misma etapa `topics_identified`), Metric Calculation (`POST /metricas/calcular` y `POST /metricas/recalcular` en MetricProcessor), Dashboard Backend (lecturas `GET /metricas/*`, `POST /metricas/recalcular-semestre` que orquesta recálculos vía MetricProcessor, `GET /metricas/archivos/entrada`, `POST /metricas/archivos/upload` y Reviews Explorer) y el listado filtrado de reviews.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Tesis OLH",
      "email": "admin@olh.com"
    }
  },
  "servers": [
    {
      "url": "https://api.olh.com",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Desarrollo local (genérico)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Dashboard Backend (local): GET /api/v1/metricas/*, POST /api/v1/metricas/recalcular-semestre, GET /api/v1/metricas/archivos/entrada y POST /api/v1/metricas/archivos/upload"
    },
    {
      "url": "http://localhost:5004",
...
    },
    {
      "url": "http://localhost:5003",
...
    },
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/verificacion/ejecutar": {
      "post": {
        "tags": [
          "Verificacion"
        ],
        "summary": "Ejecutar verificación y limpieza de un archivo",
        "description": "Disparado por el Scheduler. Descarga el blob de origen en Azure Blob Storage, valida estructura, ejecuta el pipeline de limpieza, opcionalmente sube el CSV limpio y vuelca las reviews en la tabla `reviews` (incluyendo el campo `plataforma` indicado en el body en todas las filas insertadas). Registra y actualiza `log_archivos`.",
        "operationId": "ejecutarVerificacion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificacionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Archivo verificado, limpiado y reviews cargadas correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VerificacionResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "El archivo ya fue procesado anteriormente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "ARCHIVO_YA_PROCESADO",
                    "mensaje": "El archivo reviews_olh_nov2025.csv ya fue procesado. Estado actual: completed"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Estructura del CSV inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "ESTRUCTURA_INVALIDA",
                    "mensaje": "Columnas requeridas ausentes: ['fecha_review', 'texto']"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/verificacion/archivos/{archivo_id}": {
      "get": {
        "tags": [
          "Verificacion"
        ],
        "summary": "Consultar estado de un archivo en el pipeline",
        "description": "Devuelve el estado actual y metadatos de un archivo. Usada por Sentiment Prediction, Topic Identification y Metric Calculation antes de ejecutar su etapa.",
        "operationId": "getArchivoById",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchivoId"
          }
        ],
        "responses": {
          "200": {
            "description": "Metadatos y estado del archivo",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ArchivoDetalle"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/verificacion/archivos": {
      "get": {
        "tags": [
          "Verificacion"
        ],
        "summary": "Listar archivos por estado",
        "description": "Devuelve los archivos en un estado determinado del pipeline. Usada por Sentiment Prediction para descubrir qué archivos están listos para procesar.",
        "operationId": "listarArchivos",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filtro por hotel"
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "received",
                "validated",
                "cleaned",
                "predicted",
                "topics_identified",
                "completed",
                "error"
              ]
            },
            "description": "Estado del pipeline a filtrar"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de archivos en el estado indicado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ArchivoResumen"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ArchivoId": {
        "name": "archivo_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        },
        "description": "ID del archivo en log_archivos"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente o inválido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "NO_AUTORIZADO",
                "mensaje": "Token de autenticación inválido o expirado"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "ARCHIVO_NO_ENCONTRADO",
                "mensaje": "No se encontró el recurso solicitado"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EnvelopeError"
            },
            "example": {
              "ok": false,
              "data": null,
              "error": {
                "codigo": "ERROR_INTERNO",
                "mensaje": "Error inesperado en el procesamiento"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        }
      },
      "EnvelopeError": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "error": {
            "type": "object",
            "required": [
              "codigo",
              "mensaje"
            ],
            "properties": {
              "codigo": {
                "type": "string",
                "example": "ARCHIVO_NO_ENCONTRADO"
              },
              "mensaje": {
                "type": "string",
                "example": "No se encontró el recurso solicitado"
              }
            }
          }
        }
      },
...
      "VerificacionRequest": {
        "type": "object",
        "required": [
          "hotel_id",
          "drive_id_origen",
          "nombre_archivo_origen",
          "plataforma"
        ],
        "properties": {
          "hotel_id": {
            "type": "integer",
            "example": 1
          },
          "drive_id_origen": {
            "type": "string",
            "description": "Ruta del blob de entrada dentro del contenedor Azure configurado (el nombre del campo es histórico).",
            "example": "entrada/reviews_olh_nov2025.csv"
          },
          "nombre_archivo_origen": {
            "type": "string",
            "example": "reviews_olh_nov2025.csv"
          },
          "plataforma": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "description": "Origen de las reseñas (p. ej. booking, tripadvisor). Se persiste en la columna `reviews.plataforma` para todas las filas del lote; tiene prioridad sobre una columna homónima en el CSV si existiera.",
            "example": "booking"
          }
        }
      },
      "VerificacionResponse": {
        "type": "object",
        "properties": {
          "archivo_id": {
            "type": "integer",
            "format": "int64",
            "example": 12
          },
          "nombre_archivo_origen": {
            "type": "string",
            "example": "reviews_olh_nov2025.csv"
          },
          "estado": {
            "type": "string",
            "example": "cleaned"
          },
          "total_registros": {
            "type": "integer",
            "example": 1500
          },
          "registros_validos": {
            "type": "integer",
            "example": 1432
          },
          "registros_descartados": {
            "type": "integer",
            "example": 68
          },
          "nombre_archivo_limpio": {
            "type": "string",
            "example": "clean_reviews_olh_2025-11-15.csv"
          },
          "drive_id_limpio": {
            "type": "string",
            "description": "Ruta del blob del CSV limpio en el contenedor Azure (vacío si no se subió).",
            "example": "limpios/reviews_olh_nov2025_clean.csv"
          },
          "hash": {
            "type": "string",
            "description": "SHA-256 en hexadecimal del contenido binario del blob de origen descargado desde Azure Blob.",
            "pattern": "^[a-f0-9]{64}$",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          }
        }
      },
      "ArchivoDetalle": {
        "type": "object",
        "properties": {
          "archivo_id": {
            "type": "integer",
            "format": "int64",
            "example": 12
          },
          "hotel_id": {
            "type": "integer",
            "example": 1
          },
          "nombre_archivo_origen": {
            "type": "string",
            "example": "reviews_olh_nov2025.csv"
          },
          "nombre_archivo_limpio": {
            "type": "string",
            "example": "clean_reviews_olh_2025-11-15.csv"
          },
          "drive_id_limpio": {
            "type": "string",
            "description": "Ruta del blob del CSV limpio en el contenedor Azure.",
            "example": "limpios/reviews_olh_nov2025_clean.csv"
          },
          "hash": {
            "type": "string",
            "nullable": true,
            "description": "SHA-256 en hexadecimal del blob de origen; null en registros anteriores a la migración.",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          },
          "estado": {
            "type": "string",
            "enum": [
              "received",
              "validated",
              "cleaned",
              "predicted",
              "topics_identified",
              "completed",
              "error"
            ],
            "example": "cleaned"
          },
          "total_registros": {
            "type": "integer",
            "example": 1500
          },
          "registros_validos": {
            "type": "integer",
            "example": 1432
          },
          "fecha_limpieza": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-15T08:32:00Z"
          }
        }
      },
      "ArchivoResumen": {
        "type": "object",
        "properties": {
          "archivo_id": {
            "type": "integer",
            "format": "int64",
            "example": 12
          },
          "hotel_id": {
            "type": "integer",
            "example": 1
          },
          "nombre_archivo_limpio": {
            "type": "string",
            "example": "clean_reviews_olh_2025-11-15.csv"
          },
          "estado": {
            "type": "string",
            "example": "cleaned"
          },
          "registros_validos": {
            "type": "integer",
            "example": 1432
          },
          "fecha_limpieza": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-15T08:32:00Z"
          }
        }
      },
...
    }
  }
}